chore: update bug recreation template
