+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -7,6 +7,13 @@ assignees: ''
 
 ---
 
+<!--- If you can, please recreate the bug in a quick test:
+
+Simply clone master and use our bug recreation template for chai-openapi-response-validator (https://github.com/RuntimeTools/OpenAPIValidators/blob/master/packages/chai-openapi-response-validator/test/bugRecreationTemplate.test.js)
+
+or jest-openapi (https://github.com/RuntimeTools/OpenAPIValidators/blob/master/packages/jest-openapi/__test__/bugRecreationTemplate.test.js)
+ -->
+
 **Which package are you using**?
 `chai-openapi-response-validator` / `jest-openapi`
