+++ commonTestResources/exampleOpenApiFiles/valid/bugRecreationTemplate/openapi.yml
@@ -1,10 +1,10 @@
 openapi: 3.0.0
 info:
   title: Example OpenApi 3 spec
-  description: Has various paths with responses to use in testing
+  description: Use to recreate a bug
   version: 0.1.0
 paths:
-  /test:
+  /recreate/bug:
     get:
       responses:
         '200':
@@ -12,13 +12,7 @@ paths:
           content:
             application/json:
               schema:
-                allOf:
-                  - type: object
-                    properties:
-                      expectedProperty1:
-                        type: string
-                  - type: object
-                    properties:
-                      expectedProperty2:
-                        type: string
-                    # additionalProperties: false # Uncommenting this line exposes the bug
+                type: object
+                properties:
+                  expectedProperty1:
+                    type: string
